feat: id 파라매터를 long 으로 변경

+++ src/main/java/roomescape/controller/ReservationController.java
@@ -32,7 +32,7 @@ public class ReservationController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteReservation(@PathVariable int id) {
+    public ResponseEntity<Void> deleteReservation(@PathVariable long id) {
         reservationService.deleteReservation(id);
         return ResponseEntity.noContent().build();
     }
